+++ ui/about/CountryList.tsx
@@ -1,31 +1,14 @@
 import { memo, useEffect, useState } from 'react';
 
-import {
-  AnimatePresence, LayoutGroup, motion, Variants,
-} from 'framer-motion';
+import { AnimatePresence, LayoutGroup, motion } from 'framer-motion';
 import { Country } from 'lib/model/country';
 import styled from 'styled-components';
+
+import CountryItem from './CountryItem';
 
 type Props = {
   keyword: string;
   countries: Country[];
-};
-
-const countryItemVariants: Variants = {
-  closed: {
-    opacity: 0,
-  },
-  open: {
-    opacity: 1,
-    transition: {
-      duration: 0.5,
-      delay: 0.2,
-      ease: [0, 0.71, 0.2, 1.01],
-    },
-  },
-  exit: {
-    opacity: 0,
-  },
 };
 
 function CountryList({ keyword, countries }: Props) {
@@ -53,18 +36,10 @@ function CountryList({ keyword, countries }: Props) {
         <AnimatePresence>
           {state.map(({ code, koreanName, emoji }) => (
             <CountryItem
-              role="button"
-              tabIndex={0}
               key={code}
-              layout
-              variants={countryItemVariants}
-              initial="closed"
-              animate="open"
-              exit="exit"
-            >
-              <div>{emoji}</div>
-              <div>{koreanName}</div>
-            </CountryItem>
+              emoji={emoji}
+              koreanName={koreanName}
+            />
           ))}
         </AnimatePresence>
       </CountryListWrapper>
@@ -79,27 +54,4 @@ const CountryListWrapper = styled(motion.ul)`
   list-style: none;
   margin: 0;
   padding: 0;
-`;
-
-const CountryItem = styled(motion.li)`
-  cursor: pointer;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 12px 16px;
-  background: ${({ theme }) => theme.white};
-  border-bottom: 1px solid ${({ theme }) => theme.black};
-  gap: 16px;
-
-  /* TODO - 폰트 정의 */
-  font-weight: 500;
-  font-size: 24px;
-  line-height: 32px;
-  transition: background-color 0.1s ease-in-out;
-
-  @media(hover: hover) and (pointer: fine) {
-    &:hover {
-      background-color: ${({ theme }) => theme.gray100};
-    }
-  }
 `;
