+++ ui/about/SearchCountryHeader.tsx
@@ -5,8 +5,6 @@ import { ReactNode } from 'react';
 import styled from 'styled-components';
 import Button from 'ui/common/Button';
 
-import LogoIcon from 'lib/assets/icons/offbeat_logo_draft.svg';
-
 type Props = {
   children: ReactNode;
 };
@@ -14,9 +12,6 @@ type Props = {
 function SearchCountryHeader({ children }: Props) {
   return (
     <div>
-      <SearchCountryHeaderWrapper>
-        <LogoIcon />
-      </SearchCountryHeaderWrapper>
       {children}
       <ButtonWrapper>
         <FindLocationButton type="button">
@@ -28,15 +23,6 @@ function SearchCountryHeader({ children }: Props) {
 }
 
 export default SearchCountryHeader;
-
-const SearchCountryHeaderWrapper = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  padding: 0px 16px 24px;
-  background-color: ${({ theme }) => theme.gray100};
-  width: 100%;
-`;
 
 const FindLocationButton = styled(Button)`
   /* position: fixed;
