--- ui/about/SearchInput.tsx
@@ -1,73 +1,0 @@
-'use client';
-
-import styled from 'styled-components';
-
-import SearchSvg from 'lib/assets/icons/search.svg';
-
-type Props = {
-  value: string;
-  onChange: (value: string) => void;
-};
-
-function SearchInput({ value, onChange }: Props) {
-  return (
-    <SearchInputWrapper>
-      <SearchIcon />
-      <SearchCountryInput
-        type="text"
-        placeholder="찾는 장소가 어떤 나라인가요?"
-        value={value}
-        onChange={(e) => onChange(e.target.value)}
-      />
-    </SearchInputWrapper>
-  );
-}
-
-export default SearchInput;
-
-const SearchInputWrapper = styled.div`
-  position: relative;
-
-  &:focus-within {
-    & > svg > path {
-      fill: ${({ theme }) => theme.purple500};
-    }
-  }
-`;
-
-const SearchIcon = styled(SearchSvg)`
-  cursor: pointer;
-  position: absolute;
-  top: 16px;
-  left: 16px;
-`;
-
-const SearchCountryInput = styled.input`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  height: 56px;
-  background-color: ${({ theme }) => theme.white};
-  width: 100%;
-  border-width: 1px 0px;
-  border-style: solid;
-  border-radius: 0;
-  border-color: ${({ theme }) => theme.black};
-  padding: 12px 16px 12px 56px;
-
-  /* TODO - 추후 공통 변수로 적용 */
-  font-weight: 500;
-  font-size: 24px;
-  line-height: 32px;
-  letter-spacing: -0.02em;
-  outline: none;
-
-  &::placeholder {
-    color: ${({ theme }) => theme.gray400};
-  }
-
-  &:focus {
-    caret-color: ${({ theme }) => theme.purple500};
-    border-width: 2px 0px;
-  }
-`;
